Assessment Styles Revamp: Part 4 (#918)
* #780 part 4
* Update free text answers to use textarea component
* Update associated tests

Note: There is some funkiness going on with the resize handle in Chrome that is non-existent in Firefox. Current and previous implementation just fixed height at 50px, but longer-term may want to either allow resizing, or dynamically resize the textarea to be the height of the content. In addition, right now hitting the server every time a letter is typed, which may be slow/undesirable in practice.