Link back to 360 if no automated assessments (#978)
* closes #976 
* If no automated assessments, when viewing manual assessment back link goes to 360 profile with same logic as from 360
* Update associated test

Note: Longer term I think we will want to refactor the risk area assessment component, as making unnecessary calls to load especially now that we are getting the full risk area group for patient to determine if there are "sibling" automated assessments or not. This fixes the issue short term, but longer term would like to give the assessment view some love.